VIH-6177 update events regarding judge
