--- VideoWeb/VideoWeb.EventHub/Handlers/JudgeAvailableEventHandler.cs
@@ -1,30 +1,0 @@
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.SignalR;
-using Microsoft.Extensions.Logging;
-using VideoWeb.Common.Caching;
-using VideoWeb.EventHub.Handlers.Core;
-using VideoWeb.EventHub.Hub;
-using VideoWeb.EventHub.Models;
-using VideoWeb.Services.Video;
-using EventType = VideoWeb.EventHub.Enums.EventType;
-using ParticipantState = VideoWeb.EventHub.Enums.ParticipantState;
-
-namespace VideoWeb.EventHub.Handlers
-{
-    public class JudgeAvailableEventHandler : EventHandlerBase
-    {
-        public JudgeAvailableEventHandler(IHubContext<Hub.EventHub, IEventHubClient> hubContext,
-            IConferenceCache conferenceCache, ILogger<EventHandlerBase> logger, IVideoApiClient videoApiClient) : base(
-            hubContext, conferenceCache, logger, videoApiClient)
-        {
-        }
-
-        public override EventType EventType => EventType.JudgeAvailable;
-
-        protected override async Task PublishStatusAsync(CallbackEvent callbackEvent)
-        {
-            var participantState = ParticipantState.Available;
-            await PublishParticipantStatusMessage(participantState);
-        }
-    }
-}
