+++ VideoWeb/VideoWeb.UnitTests/Builders/EventComponentHelper.cs
@@ -54,10 +54,6 @@ namespace VideoWeb.UnitTests.Builders
                 new HelpEventHandler(eventHubContextMock.Object, ConferenceCache, logger.Object, apiClientMock.Object),
                 new JoinedEventHandler(eventHubContextMock.Object, ConferenceCache, logger.Object,
                     apiClientMock.Object),
-                new JudgeAvailableEventHandler(eventHubContextMock.Object, ConferenceCache, logger.Object,
-                    apiClientMock.Object),
-                new JudgeUnavailableEventHandler(eventHubContextMock.Object, ConferenceCache, logger.Object,
-                    apiClientMock.Object),
                 new LeaveEventHandler(eventHubContextMock.Object, ConferenceCache, logger.Object, apiClientMock.Object),
                 new PauseEventHandler(eventHubContextMock.Object, ConferenceCache, logger.Object, apiClientMock.Object),
                 new SuspendEventHandler(eventHubContextMock.Object, ConferenceCache, logger.Object,
