+++ VideoWeb/VideoWeb.UnitTests/EventHandlers/EventHandlerFactoryTests.cs
@@ -18,8 +18,6 @@ namespace VideoWeb.UnitTests.EventHandlers
         [TestCase(EventType.Transfer, typeof(TransferEventHandler))]
         [TestCase(EventType.ParticipantJoining, typeof(ParticipantJoiningEventHandler))]
         [TestCase(EventType.VhoCall, typeof(VhOfficerCallEventHandler))]
-        [TestCase(EventType.JudgeAvailable, typeof(JudgeAvailableEventHandler))]
-        [TestCase(EventType.JudgeUnavailable, typeof(JudgeUnavailableEventHandler))]
         public void Should_return_instance_of_event_handler_when_factory_get_is_called_with_valid_request(
             EventType eventType, Type typeOfEventHandler)
         {
